two more downcasts

+++ khmer/_khmermodule.cc
@@ -2287,7 +2287,7 @@ static PyObject * hashbits_repartition_largest_partition(PyObject * self, PyObje
 
     CountingHash * counting = counting_o->counting;
 
-    unsigned int next_largest = subset_p->repartition_largest_partition(distance,
+    unsigned long next_largest = subset_p->repartition_largest_partition(distance,
                                 threshold, frequency, *counting);
 
     return PyInt_FromLong(next_largest);
